Frontend structure refactored
