+++ frontend/src/components/Landing/index.ts
@@ -1,0 +1,3 @@
+import LandingPage from './LandingPage';
+
+export default LandingPage;
