+++ frontend/src/components/Recipes/Recipes.tsx
@@ -1,5 +1,6 @@
+import React from 'react';
 
-export default function RecipesPage() {
+export default function Recipes() {
   return (
     <div>
       <h1 className="text-2xl font-bold">Recipes</h1>
