+++ frontend/src/index.tsx
@@ -3,22 +3,27 @@ import * as ReactDOM from 'react-dom/client';
 import { createHashRouter, RouterProvider } from 'react-router-dom';
 import './index.css';
 
-import LandingPage from './routes/page';
-import RecipesPage from './routes/recipes/page';
+import LandingPage from './components/Landing';
+import RecipesPage from './components/Recipes/Recipes'
+import Test from './components/Tests/Test';
 
 const router = createHashRouter([
-  {
-    path: '/',
-    Component: LandingPage,
-  },
-  {
-    path: '/recipes',
-    Component: RecipesPage
-  }
+	{
+		path: '/',
+		Component: LandingPage,
+	},
+	{
+		path: '/recipes',
+		Component: RecipesPage,
+	},
+	{
+		path: '/test',
+		Component: Test,
+	},
 ]);
 
 ReactDOM.createRoot(document.getElementById('root') as Element).render(
-  <React.StrictMode>
-    <RouterProvider router={router} />
-  </React.StrictMode>
+	<React.StrictMode>
+		<RouterProvider router={router} />
+	</React.StrictMode>
 );
